rename add-account to account
