+++ src/components/account-dialog/account-dialog.scss
@@ -1,4 +1,4 @@
-.add-account-dialog {
+.account-dialog {
   &__select-menu {
     top: inherit !important;
     left: inherit !important;
