+++ src/index.scss
@@ -24,7 +24,7 @@
 @import "./components/toolbar/toolbar";
 @import "./components/transactions/transactions";
 @import "./components/transaction-dialog/transaction-dialog";
-@import "./components/add-account-dialog/add-account-dialog";
+@import "./components/account-dialog/account-dialog";
 @import "./components/button-more/button-more";
 @import "./components/table-transactions/table-transactions";
 @import "./components/chart/chart";
